Created VRDoodler class with callback functions for better scalability.
Basically you can define custom onClick behavior outside of the VRDoodler class,
and can create an instance of VRDoodler wherever you wish. Perhaps an idea for
the future is to be able to specify /where/ exactly you want the instance to be
created on the screen, take as inputs to the constructor custom callbacks, etc.

Another benefit of this is that you can now put whatever else you like in the
animate() loop in the .html file - VRDoodler has its own animate callback
within the object.